refactor: extract common code of preparing pipeline
