+++ cli/lib/log.ts
@@ -1,0 +1,4 @@
+import debug from 'debug'
+
+export const log = debug('cube-creator')
+log.enabled = false
