+++ cli/test/lib/job.test.ts
@@ -5,7 +5,7 @@ import env from '@cube-creator/core/env'
 import { rdfs, schema } from '@tpluscode/rdf-ns-builders'
 import $rdf from 'rdf-ext'
 import { JobIterator } from '../../lib/job'
-import { insertTestProject } from '@cube-creator/testing/lib'
+import { insertTestProject } from '@cube-creator/testing/lib/seedData'
 import { setupEnv } from '../support/env'
 import { Table } from '@rdfine/csvw'
 import { log } from '../support/logger'
