+++ cli/test/lib/output-mapper.test.ts
@@ -3,7 +3,7 @@ import { expect } from 'chai'
 import $rdf from 'rdf-ext'
 import { cube } from '@cube-creator/core/namespace'
 import { mapDimensions } from '../../lib/output-mapper'
-import { insertTestProject } from '@cube-creator/testing/lib'
+import { insertTestProject } from '@cube-creator/testing/lib/seedData'
 import { setupEnv } from '../support/env'
 import { env } from 'process'
